Make component for page link
