+++ src/components/TrackTool.tsx
@@ -1,17 +1,11 @@
 import React from 'react'
-import {
-  Link,
-  Switch,
-  Route,
-  Redirect,
-  useRouteMatch,
-  useLocation,
-  generatePath,
-} from 'react-router-dom'
+import { Link, Switch, Route, Redirect, useRouteMatch } from 'react-router-dom'
 
 import { TrackContributing } from './TrackContributing'
 import { TrackMaintaining } from './TrackMaintaining'
 import { TrackNewExercise } from './TrackNewExercise'
+import { PageLink } from './PageLink'
+
 import {
   ProvideActionable,
   useProvideActionableState,
@@ -42,15 +36,9 @@ function TogglePageButton(): JSX.Element {
 
   return (
     <div className="btn-group">
-      <TrackToolPageLink to={`${match.url}/contributing`}>
-        Contributing
-      </TrackToolPageLink>
-      <TrackToolPageLink to={`${match.url}/maintaining`}>
-        Maintaining
-      </TrackToolPageLink>
-      <TrackToolPageLink to={`${match.url}/new-exercise`}>
-        New exercise
-      </TrackToolPageLink>
+      <PageLink to={`${match.url}/contributing`}>Contributing</PageLink>
+      <PageLink to={`${match.url}/maintaining`}>Maintaining</PageLink>
+      <PageLink to={`${match.url}/new-exercise`}>New exercise</PageLink>
     </div>
   )
 }
@@ -76,29 +64,5 @@ function TrackToolPage(): JSX.Element {
       <Route path={`${match.path}/new-exercise`} component={TrackNewExercise} />
       <Redirect to={`${match.path}/contributing`} />
     </Switch>
-  )
-}
-
-interface TrackToolPageLinkProps {
-  to: string
-  children: React.ReactNode
-}
-
-function TrackToolPageLink({
-  to,
-  children,
-}: TrackToolPageLinkProps): JSX.Element {
-  const location = useLocation()
-  const match = useRouteMatch()
-  const path = generatePath(to, match.params)
-  const active = location.pathname.startsWith(path)
-
-  return (
-    <Link
-      to={to}
-      className={`btn btn-sm btn-outline-primary ${active ? 'active' : ''}`}
-    >
-      {children}
-    </Link>
   )
 }
